requirements.txt
python-docx>=1.1.0
PySide6>=6.6.0
lxml>=5.0.0
latex2mathml>=3.79.0
olefile>=0.47
pywin32>=306; platform_system == "Windows"
pytest>=8.0.0
